Refactor callback system (#7)
* Before testing all the commands (GetControllersIds: OK)

* Fix bugs (shared_from_this(), ErrorResponse not handled,...

* Fix bugs after tests with all the commands, with new callback system

* Clean includes (makes Cmake faster ???)

* Merge manufacturer data

* Add --logging option to set logging level on runtime

* Use ioctl to get HCI controllers ids at the startup

* Use string address instead of bytes array in AddDevice,RemoveDevice and Disconnect

* Remove header files from SOURCE_FILES cmake variable (useless...)

* Remove nested namespace to be gcc5 compatible

* Move AbstractExtractor templates to cpp file

* Replace MGMT scan by HCI scan to have a scan forever command

* Fix bugs with Meta packet

+++ platforms/linux/src/Format/HCI/HCIFormatBuilder.hpp
@@ -1,15 +1,11 @@
 #ifndef BABLE_LINUX_HCIFORMATBUILDER_HPP
 #define BABLE_LINUX_HCIFORMATBUILDER_HPP
 
+#include "./constants.hpp"
+
 #if __BYTE_ORDER__ == __ORDER_PDP_ENDIAN__
 #error "Byte order not suported (PDP endian)"
 #endif
-
-#include <array>
-#include <cstdint>
-#include <vector>
-#include <string>
-#include "./constants.hpp"
 
 class HCIFormatBuilder {
 
@@ -18,7 +14,7 @@ public:
   explicit HCIFormatBuilder(uint16_t controller_id);
 
   // Setters
-  HCIFormatBuilder& set_opcode(uint8_t code);
+  HCIFormatBuilder& set_opcode(uint16_t code);
   HCIFormatBuilder& set_controller_id(uint16_t controller_id);
   HCIFormatBuilder& set_connection_handle(uint16_t connection_handle);
 
@@ -41,7 +37,7 @@ private:
   std::vector<uint8_t> generate_command_header();
   std::vector<uint8_t> generate_async_data_header();
 
-  uint8_t m_opcode;
+  uint16_t m_opcode;
   uint16_t m_controller_id;
   uint16_t m_connection_handle;
   std::vector<uint8_t> m_formatted_data;
